Rename provider for ngx-flagr/core
