+++ projects/core/src/index.ts
@@ -2,5 +2,5 @@ export { FeatureFlag } from './feature-flag';
 export { FeatureFlagDirective } from './feature-flag.directive';
 export { FeatureFlagService } from './feature-flag.service';
 export { NgxFlagrModule } from './ngx-flagr.module';
-export { provideNgxFlagr } from './provide-ngx-flagr';
+export { provideNgxFlagr } from './provide-core';
 export { FEATURE_FLAG_SERVICE } from './tokens';
